pyproject.toml
[project]
name = "mf-automated-perception"
version = "0.1.0"
description = "CLI and runner for automated perception procedures"
requires-python = "==3.10.*"

# this has a lot dependencies for ros2 humble
dependencies = [
  "typer>=0.12",
  "rich>=13.0",
  "pydantic",
  "pytest==6.2.5",
  "rerun-sdk",
  "pyyaml",
  "tqdm",
  "opencv-python==4.7.0.72",
  "numpy==1.23.5",
  "seaborn",
  "lark>=1.1.5,<2.0"
]

[project.scripts]
mf-eye = "mf_automated_perception.host.cli.main:app"


[tool.ruff]
line-length = 88
target-version = "py311"

select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "W",  # pycodestyle warnings
  "B",  # bugbear
  "I",  # import order
]

ignore = [
  "E501",  # line-length (formatter가 책임)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]


# ============================
# pytest configuration
# ============================
[tool.pytest.ini_options]
# prevent ros2 pytest
addopts = "-ra -q "
testpaths = ["tests"]


markers = [
  "unit: fast unit tests without side effects",
  "integration: integration tests with filesystem side effects",
]


[tool.setuptools.packages.find]
where = ["."]
include = ["mf_automated_perception*"]

[tool.setuptools]
package-dir = {"" = "."}
